admin_backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
typing_extensions==4.15.0
Werkzeug==3.1.4
httpx>=0.25.0
supabase>=2.0.0
python-dotenv>=1.0.0
